デバッグ

+++ CrayfishMonitor/CrayfishMonitor-Desktop/Models/IMeasurementData.cs
@@ -5,7 +5,7 @@ namespace CrayfishMonitor_Desktop.Models
     public interface IMeasurementData
     {
         public DateTime Time { get; set; }
-        public double Elapsed { get; set; }
+        public long Elapsed { get; set; }
         public double Voltage { get; set; }
     }
 }
